Install AFNI via package manager
Previously, this led to installation errors in libgsl.so. According to Ryan,
the hacky if condition around lines 80 (now commented out) would prevent/fix
these problems on system he had tested on, but on a current Debian buster and
a Debain bullseye/sid installation I have access to this does not work.

Changing the Dockerfile is problematic and defies the very idea of a Dockerfile,
but atm I wouldn't be able to successfully preprocess otherwise. This approach
however will need checks whether Images from old and changed Dockerfile result
in the same outputs.